Refactor: extract sanitizeSchema, add $defs support and missing tests
- Move sanitizeSchema to a dedicated internal module (not publicly exported)
- Add $defs handling alongside definitions for newer JSON Schema drafts
- Add missing tests for allOf, not, patternProperties, additionalProperties
  boolean passthrough, and $defs sanitization
